Add ProductCounter component for increment value in basket & small changes in ProcutListComponent
