+++ .vscode/settings.json
@@ -1,17 +1,16 @@
 {
-   "editor.formatOnSave": true,
-   "typescript.tsdk": "node_modules/typescript/lib",
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.codeActionsOnSave": {
-        "source.fixAll": true
-    },
-    "graphql-config.dotEnvPath": "./.env",
-    "graphql-config.load.configName": "graphql",
-    "files.associations": {
-        "*.css":"tailwindcss",
-    },
-"editor.quickSuggestions": {
- 
-    "strings": true
-}
+	"editor.formatOnSave": true,
+	"typescript.tsdk": "node_modules/typescript/lib",
+	"editor.defaultFormatter": "esbenp.prettier-vscode",
+	"editor.codeActionsOnSave": {
+		"source.fixAll": true
+	},
+	"graphql-config.dotEnvPath": "./.env",
+	"graphql-config.load.configName": "graphql",
+	"files.associations": {
+		"*.css": "tailwindcss"
+	},
+	"editor.quickSuggestions": {
+		"strings": true
+	}
 }
