+++ package.json
@@ -19,6 +19,7 @@
 		"postcss": "8.4.29",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
+		"server-only": "^0.0.1",
 		"tailwindcss": "3.3.3",
 		"typescript": "5.2.2"
 	},
